# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Payment gateway aggregator — VA, QRIS, and disbursement via Midtrans, Xendit, DOKU, iPaymu.\n\n## Authentication\nMerchant endpoints (/v1/*): `X-API-Key` header with your API key (`wpay_live_*` or `wpay_test_*`).\nAdmin endpoints (/admin/*): `Authorization: Bearer \u003caccess_token\u003e` from POST /admin/login.\n\n## Idempotency\nPayment and disbursement creation support `Idempotency-Key` header. Use unique key per request to safely retry.\n\n## Webhook\nWanpey signs outbound webhooks with `X-Wanpey-Signature` (HMAC-SHA256). Verify using your webhook secret.",
        "title": "Wanpey Payment Gateway API",
        "contact": {
            "name": "Wanpey Support",
            "email": "support@wanpey.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/payments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated list of payments for the authenticated merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "List payments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status: pending, paid, expired, cancelled, failed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by provider: midtrans, xendit, doku, ipaymu",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by method: va, qris",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wanpey_core_internal_usecase.PaymentOutput"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/qris": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a QRIS payment and returns a QR code string for the customer to scan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create QRIS payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique key to prevent duplicate payments on retry",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "QRIS payment request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_internal_usecase.CreateQRISInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.PaymentOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/va": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a VA payment via the specified provider. Returns VA number and bank details for the customer to pay.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create Virtual Account payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique key to prevent duplicate payments on retry",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "VA payment request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_internal_usecase.CreateVAInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.PaymentOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns full payment detail including status, VA/QR data, and fee breakdown.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.PaymentOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancels a pending payment. Uses a two-step process (pending → cancelling → cancelled) for financial safety. Only pending payments can be cancelled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Cancel a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment cancelled successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Payment is already in a final state",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "wanpey_core_internal_domain_entity.BankCode": {
            "type": "string",
            "enum": [
                "BCA",
                "BNI",
                "BRI",
                "BSI",
                "MANDIRI",
                "PERMATA",
                "CIMB"
            ],
            "x-enum-varnames": [
                "BankBCA",
                "BankBNI",
                "BankBRI",
                "BankBSI",
                "BankMandiri",
                "BankPermata",
                "BankCIMB"
            ]
        },
        "wanpey_core_internal_domain_entity.Currency": {
            "type": "string",
            "enum": [
                "IDR"
            ],
            "x-enum-varnames": [
                "CurrencyIDR"
            ]
        },
...
        "wanpey_core_internal_domain_entity.PaymentMethod": {
            "type": "string",
            "enum": [
                "va",
                "qris",
                "cc"
            ],
            "x-enum-comments": {
                "PaymentMethodCC": "coming soon"
            },
            "x-enum-descriptions": [
                "",
                "",
                "coming soon"
            ],
            "x-enum-varnames": [
                "PaymentMethodVA",
                "PaymentMethodQRIS",
                "PaymentMethodCC"
            ]
        },
        "wanpey_core_internal_domain_entity.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "cancelling",
                "paid",
                "expired",
                "failed",
                "cancelled"
            ],
            "x-enum-comments": {
                "PaymentStatusCancelling": "transient: provider call in-flight"
            },
            "x-enum-descriptions": [
                "",
                "transient: provider call in-flight",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusCancelling",
                "PaymentStatusPaid",
                "PaymentStatusExpired",
                "PaymentStatusFailed",
                "PaymentStatusCancelled"
            ]
        },
...
        "wanpey_core_internal_domain_entity.Provider": {
            "type": "string",
            "enum": [
                "midtrans",
                "xendit",
                "doku",
                "ipaymu"
            ],
            "x-enum-varnames": [
                "ProviderMidtrans",
                "ProviderXendit",
                "ProviderDoku",
                "ProviderIPaymu"
            ]
        },
...
        "wanpey_core_internal_usecase.CreateQRISInput": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "customer_email",
                "customer_name",
                "customer_phone",
                "expiry_at",
                "provider"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "callback_url": {
                    "type": "string",
                    "maxLength": 500
                },
                "currency": {
                    "enum": [
                        "IDR"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                        }
                    ]
                },
                "customer_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "customer_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "customer_phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "expiry_at": {
                    "type": "string"
                },
                "provider": {
                    "enum": [
                        "midtrans",
                        "xendit",
                        "doku",
                        "ipaymu"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Provider"
                        }
                    ]
                }
            }
        },
        "wanpey_core_internal_usecase.CreateVAInput": {
            "type": "object",
            "required": [
                "amount",
                "bank_code",
                "currency",
                "customer_email",
                "customer_name",
                "customer_phone",
                "expiry_at",
                "provider"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bank_code": {
                    "enum": [
                        "BCA",
                        "BNI",
                        "BRI",
                        "BSI",
                        "MANDIRI",
                        "PERMATA",
                        "CIMB"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.BankCode"
                        }
                    ]
                },
                "callback_url": {
                    "type": "string",
                    "maxLength": 500
                },
                "currency": {
                    "enum": [
                        "IDR"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                        }
                    ]
                },
                "customer_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "customer_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "customer_phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "expiry_at": {
                    "type": "string"
                },
                "provider": {
                    "enum": [
                        "midtrans",
                        "xendit",
                        "doku",
                        "ipaymu"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Provider"
                        }
                    ]
                }
            }
        },
...
        "wanpey_core_internal_usecase.PaymentOutput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bank_code": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.BankCode"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "expiry_at": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "fee_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.PaymentMethod"
                },
                "paid_at": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.Provider"
                },
                "qr_image_url": {
                    "type": "string"
                },
                "qr_string": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.PaymentStatus"
                },
                "va_number": {
                    "type": "string"
                }
            }
        },
...
        "wanpey_core_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wanpey_core_pkg_response.FieldDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "wanpey_core_pkg_response.FieldDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "wanpey_core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Pagination"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "wanpey_core_pkg_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT — Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Merchant API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}